refactor(editor): Move `settings` api to `@n8n/rest-api-client` (no-changelog) (#16186)
