+++ packages/@n8n/constants/src/api.ts
@@ -1,0 +1,1 @@
+export const N8N_IO_BASE_URL = 'https://api.n8n.io/api/';
