+++ packages/@n8n/constants/src/community-nodes.ts
@@ -1,0 +1,1 @@
+export const NPM_COMMUNITY_NODE_SEARCH_API_URL = 'https://api.npms.io/v2/';
