+++ packages/@n8n/constants/src/index.ts
@@ -1,4 +1,6 @@
+export * from './api';
 export * from './browser';
+export * from './community-nodes';
 export * from './instance';
 
 export const LICENSE_FEATURES = {
