+++ packages/frontend/@n8n/rest-api-client/src/api/communityNodes.ts
@@ -1,3 +1,4 @@
+import { NPM_COMMUNITY_NODE_SEARCH_API_URL } from '@n8n/constants';
 import type { PublicInstalledPackage } from 'n8n-workflow';
 
 import type { IRestApiContext } from '../types';
@@ -28,4 +29,13 @@ export async function updatePackage(
 	name: string,
 ): Promise<PublicInstalledPackage> {
 	return await makeRestApiRequest(context, 'PATCH', '/community-packages', { name });
+}
+
+export async function getAvailableCommunityPackageCount(): Promise<number> {
+	const response = await get(
+		NPM_COMMUNITY_NODE_SEARCH_API_URL,
+		'search?q=keywords:n8n-community-node-package',
+	);
+
+	return response.total || 0;
 }
