+++ packages/frontend/@n8n/rest-api-client/src/api/index.ts
@@ -9,7 +9,9 @@ export * from './mfa';
 export * from './nodeTypes';
 export * from './npsSurvey';
 export * from './orchestration';
+export * from './prompts';
 export * from './roles';
+export * from './settings';
 export * from './sso';
 export * from './ui';
 export * from './versions';
