+++ packages/frontend/@n8n/rest-api-client/src/api/settings.ts
@@ -1,0 +1,8 @@
+import type { FrontendSettings } from '@n8n/api-types';
+
+import type { IRestApiContext } from '../types';
+import { makeRestApiRequest } from '../utils';
+
+export async function getSettings(context: IRestApiContext): Promise<FrontendSettings> {
+	return await makeRestApiRequest(context, 'GET', '/settings');
+}
