+++ packages/frontend/editor-ui/src/stores/settings.store.test.ts
@@ -13,7 +13,7 @@ const { sessionStarted } = vi.hoisted(() => ({
 	sessionStarted: vi.fn(),
 }));
 
-vi.mock('@/api/settings', () => ({
+vi.mock('@n8n/rest-api-client/api/settings', () => ({
 	getSettings,
 }));
